[FIX] Weather Type 변경

+++ app/src/main/java/com/doyeon/weatherapp/data/model/WeatherData.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 data class WeatherData(
     @SerializedName("weather_state_name") val weatherName: String,
     @SerializedName("weather_state_abbr") val weatherSummary: String,
-    @SerializedName("the_temp") val temperature: Int,
-    @SerializedName("humidity") val humidity: Int,
+    @SerializedName("the_temp") val temperature: Float,
+    @SerializedName("humidity") val humidity: Float,
 )
